refactor(Domain): 패키지 구조 변경
Lotto와 LottoMachine을 Domain 패키지로 옮김
